dodanie klasy DataContract i przeniesienie do niej klasy z properties o atrybutach DataContract ; poprawa atrybutów WebInvoke dla metod w interfejsie IRESTService ; dodanie metod GetDataJSON i GetDataXML do klasy serwisu RESTService
