+++ RESTClient/RESTClient.csproj
@@ -92,6 +92,7 @@
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DataContract.cs" />
     <Compile Include="RESTService.svc.cs">
       <DependentUpon>RESTService.svc</DependentUpon>
     </Compile>
